Revert "feat(bibframe-search): Implement bibframe search functionality"
This reverts commit c01bc919f3516b1af2d3b0c69b5f633aa9e6b7d9.
