+++ src/main/java/org/folio/search/cql/LccnSearchTermProcessor.java
@@ -1,18 +1,15 @@
 package org.folio.search.cql;
 
 import lombok.RequiredArgsConstructor;
-import org.folio.search.service.lccn.LccnNormalizer;
+import org.folio.search.utils.SearchUtils;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
 public class LccnSearchTermProcessor implements SearchTermProcessor {
 
-  private final LccnNormalizer lccnNormalizer;
-
   @Override
   public String getSearchTerm(String inputTerm) {
-    return lccnNormalizer.apply(inputTerm)
-      .orElse(null);
+    return SearchUtils.normalizeLccn(inputTerm);
   }
 }
