+++ src/main/java/org/folio/search/integration/KafkaMessageListener.java
@@ -10,7 +10,6 @@ import static org.folio.search.domain.dto.ResourceEventType.REINDEX;
 import static org.folio.search.utils.SearchConverterUtils.getEventPayload;
 import static org.folio.search.utils.SearchConverterUtils.getResourceEventId;
 import static org.folio.search.utils.SearchConverterUtils.getResourceSource;
-import static org.folio.search.utils.SearchUtils.BIBFRAME_RESOURCE;
 import static org.folio.search.utils.SearchUtils.ID_FIELD;
 import static org.folio.search.utils.SearchUtils.INSTANCE_ID_FIELD;
 import static org.folio.search.utils.SearchUtils.INSTANCE_RESOURCE;
@@ -191,22 +190,6 @@ public class KafkaMessageListener {
     var batch = consumerRecords.stream()
       .map(ConsumerRecord::value)
       .map(location -> location.id(getResourceEventId(location) + "|" + location.getTenant()))
-      .toList();
-
-    indexResources(batch, resourceService::indexResources);
-  }
-
-  @KafkaListener(
-    id = KafkaConstants.BIBFRAME_LISTENER_ID,
-    containerFactory = "standardListenerContainerFactory",
-    groupId = "#{folioKafkaProperties.listener['bibframe'].groupId}",
-    concurrency = "#{folioKafkaProperties.listener['bibframe'].concurrency}",
-    topicPattern = "#{folioKafkaProperties.listener['bibframe'].topicPattern}")
-  public void handleBibframeEvents(List<ConsumerRecord<String, ResourceEvent>> consumerRecords) {
-    log.info("Processing bibframe events from Kafka [number of events: {}]", consumerRecords.size());
-    var batch = consumerRecords.stream()
-      .map(ConsumerRecord::value)
-      .map(bibframe -> bibframe.resourceName(BIBFRAME_RESOURCE).id(getResourceEventId(bibframe)))
       .toList();
 
     indexResources(batch, resourceService::indexResources);
