--- src/main/java/org/folio/search/service/lccn/LccnNormalizer.java
@@ -1,7 +1,0 @@
-package org.folio.search.service.lccn;
-
-import java.util.Optional;
-import java.util.function.Function;
-
-public interface LccnNormalizer extends Function<String, Optional<String>> {
-}
