+++ src/main/java/org/folio/search/service/setter/AbstractLccnProcessor.java
@@ -3,28 +3,24 @@ package org.folio.search.service.setter;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import org.folio.search.domain.dto.Identifier;
 import org.folio.search.integration.ReferenceDataService;
-import org.folio.search.service.lccn.LccnNormalizer;
+import org.folio.search.utils.SearchUtils;
 
 public abstract class AbstractLccnProcessor<T> extends AbstractIdentifierProcessor<T> {
 
   private static final List<String> LCCN_IDENTIFIER_NAME = List.of("LCCN", "Canceled LCCN");
-  private final LccnNormalizer lccnNormalizer;
 
-  protected AbstractLccnProcessor(ReferenceDataService referenceDataService, LccnNormalizer lccnNormalizer) {
+  protected AbstractLccnProcessor(ReferenceDataService referenceDataService) {
     super(referenceDataService, LCCN_IDENTIFIER_NAME);
-    this.lccnNormalizer = lccnNormalizer;
   }
 
   @Override
   public Set<String> getFieldValue(T entity) {
     return filterIdentifiersValue(getIdentifiers(entity)).stream()
-      .map(lccnNormalizer)
-      .flatMap(Optional::stream)
+      .map(SearchUtils::normalizeLccn)
       .filter(Objects::nonNull)
       .collect(Collectors.toCollection(LinkedHashSet::new));
   }
