+++ src/main/java/org/folio/search/service/setter/authority/LccnAuthorityProcessor.java
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.folio.search.domain.dto.Authority;
 import org.folio.search.domain.dto.Identifier;
 import org.folio.search.integration.ReferenceDataService;
-import org.folio.search.service.lccn.LccnNormalizer;
 import org.folio.search.service.setter.AbstractLccnProcessor;
 import org.springframework.stereotype.Component;
 
@@ -17,10 +16,9 @@ public class LccnAuthorityProcessor extends AbstractLccnProcessor<Authority> {
    * Used by dependency injection.
    *
    * @param referenceDataService {@link ReferenceDataService} bean
-   * @param lccnNormalizer {@link LccnNormalizer} bean
    */
-  public LccnAuthorityProcessor(ReferenceDataService referenceDataService, LccnNormalizer lccnNormalizer) {
-    super(referenceDataService, lccnNormalizer);
+  public LccnAuthorityProcessor(ReferenceDataService referenceDataService) {
+    super(referenceDataService);
   }
 
   @Override
