+++ src/main/java/org/folio/search/service/setter/instance/LccnInstanceProcessor.java
@@ -6,7 +6,6 @@ import java.util.Optional;
 import org.folio.search.domain.dto.Identifier;
 import org.folio.search.domain.dto.Instance;
 import org.folio.search.integration.ReferenceDataService;
-import org.folio.search.service.lccn.LccnNormalizer;
 import org.folio.search.service.setter.AbstractLccnProcessor;
 import org.springframework.stereotype.Component;
 
@@ -20,10 +19,9 @@ public class LccnInstanceProcessor extends AbstractLccnProcessor<Instance> {
    * Used by dependency injection.
    *
    * @param referenceDataService {@link ReferenceDataService} bean
-   * @param lccnNormalizer {@link LccnNormalizer} bean
    */
-  public LccnInstanceProcessor(ReferenceDataService referenceDataService, LccnNormalizer lccnNormalizer) {
-    super(referenceDataService, lccnNormalizer);
+  public LccnInstanceProcessor(ReferenceDataService referenceDataService) {
+    super(referenceDataService);
   }
 
   @Override
