+++ src/main/resources/application.yml
@@ -135,10 +135,6 @@ folio:
         concurrency: ${KAFKA_LOCATION_CONCURRENCY:1}
         topic-pattern: (${folio.environment}\.)(.*\.)inventory\.location
         group-id: ${folio.environment}-mod-search-location-type-group
-      bibframe:
-        concurrency: ${KAFKA_BIBFRAME_CONCURRENCY:1}
-        topic-pattern: (${folio.environment}\.)(.*\.)search\.bibframe
-        group-id: ${folio.environment}-mod-search-bibframe-group
   okapiUrl: ${okapi.url}
   logging:
     request:
