--- src/main/resources/elasticsearch/index/bibframe.json
@@ -1,68 +1,0 @@
-{
-  "index": {
-    "number_of_shards": 4,
-    "number_of_replicas": 2,
-    "refresh_interval": "1s",
-    "codec": "best_compression",
-    "mapping.total_fields.limit": 1000
-  },
-  "analysis": {
-    "filter": {
-      "folio_word_delimiter_graph": {
-        "type": "word_delimiter_graph",
-        "catenate_words": true
-      }
-    },
-    "normalizer": {
-      "keyword_lowercase": {
-        "filter": [
-          "lowercase",
-          "trim"
-        ],
-        "type": "custom"
-      },
-      "keyword_uppercase": {
-        "filter": [
-          "uppercase",
-          "trim"
-        ],
-        "type": "custom"
-      },
-      "keyword_trimmed": {
-        "filter": [
-          "trim"
-        ],
-        "type": "custom"
-      }
-    },
-    "analyzer": {
-      "source_analyzer": {
-        "tokenizer": "icu_tokenizer",
-        "filter": [
-          "folio_word_delimiter_graph",
-          "icu_folding"
-        ],
-        "char_filter": [
-          "and_char_replacement"
-        ],
-        "type": "custom"
-      },
-      "whitespace_lowercase_analyzer": {
-        "tokenizer": "whitespace",
-        "filter": [
-          "lowercase",
-          "icu_folding"
-        ],
-        "type": "custom"
-      }
-    },
-    "tokenizers": { },
-    "char_filter": {
-      "and_char_replacement": {
-        "type": "pattern_replace",
-        "pattern": " & ",
-        "replacement": " and "
-      }
-    }
-  }
-}
