--- src/main/resources/swagger.api/examples/result/bibframeSearchResult.yaml
@@ -1,38 +1,0 @@
-value:
-  searchQuery: "query string"
-  content:
-    - id: "1"
-      titles:
-        - value: "Title Value"
-          type: "Main"
-      contributors:
-        - name: "John Doe"
-          type: "Person"
-          isCreator: true
-      languages:
-        - value: "eng"
-      classifications:
-        - number: "1234"
-          source: "ddc"
-      subjects:
-        - value: "Subject"
-      instances:
-        - id: "2"
-          titles:
-            - value: "Title Value"
-              type: "Main"
-          identifiers:
-            - value: "1234567890"
-              type: "ISBN"
-          contributors:
-            - name: "John Doe"
-              type: "Person"
-              isCreator: true
-          publications:
-            - name: "Publisher Name"
-              date: "2022"
-          editionStatements:
-            - value: "Edition 1"
-  pageNumber: 0
-  totalPages: 10
-  totalRecords: 100
