--- src/main/resources/swagger.api/parameters/bibframe-limit-param.yaml
@@ -1,8 +1,0 @@
-in: query
-name: limit
-description: Limit the number of elements returned in the response.
-schema:
-  type: integer
-  minimum: 0
-  maximum: 100
-  default: 10
