--- src/main/resources/swagger.api/schemas/response/bibframeSearchResult.yaml
@@ -1,20 +1,0 @@
-description: "Bibframe search result response"
-type: "object"
-properties:
-  searchQuery:
-    type: "string"
-    description: "Initial search query"
-  content:
-    type: "array"
-    description: "List of bibframe records found"
-    items:
-      $ref: "../../schemas/dto/bibframe/bibframe.yaml"
-  pageNumber:
-    type: "integer"
-    description: "Current results page number, 0 by default"
-  totalPages:
-    type: "integer"
-    description: "Total pages count"
-  totalRecords:
-    type: "integer"
-    description: "Total results count"
