+++ src/test/java/org/folio/search/service/setter/instance/LccnInstanceProcessorTest.java
@@ -16,15 +16,14 @@ import org.folio.search.domain.dto.Identifier;
 import org.folio.search.domain.dto.Instance;
 import org.folio.search.integration.ReferenceDataService;
 import org.folio.search.model.client.CqlQueryParam;
-import org.folio.search.service.lccn.DefaultLccnNormalizer;
 import org.folio.spring.testing.type.UnitTest;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
@@ -36,13 +35,8 @@ class LccnInstanceProcessorTest {
   @Mock
   private ReferenceDataService referenceDataService;
 
+  @InjectMocks
   private LccnInstanceProcessor lccnProcessor;
-
-  @BeforeEach
-  void setup() {
-    var lccnNormalizer = new DefaultLccnNormalizer();
-    lccnProcessor = new LccnInstanceProcessor(referenceDataService, lccnNormalizer);
-  }
 
   @MethodSource("lccnDataProvider")
   @DisplayName("getFieldValue_parameterized")
