+++ src/test/java/org/folio/search/support/base/ApiEndpoints.java
@@ -77,10 +77,6 @@ public class ApiEndpoints {
     return "/browse/classification-numbers/" + optionType.getValue() + "/instances";
   }
 
-  public static String bibframeSearchPath() {
-    return "/search/bibframe";
-  }
-
   public static String authorityBrowsePath() {
     return "/browse/authorities";
   }
