+++ src/test/java/org/folio/search/utils/TestConstants.java
@@ -37,7 +37,6 @@ public class TestConstants {
   public static final String INVENTORY_BOUND_WITH_TOPIC = "inventory.bound-with";
   public static final String INVENTORY_CLASSIFICATION_TYPE_TOPIC = "inventory.classification-type";
   public static final String CONSORTIUM_INSTANCE_TOPIC = "search.consortium.instance";
-  public static final String BIBFRAME_TOPIC = "search.bibframe";
 
   public static final String LOCAL_CN_TYPE = "6fd29f52-5c9c-44d0-b529-e9c5eb3a0aba";
   public static final String FOLIO_CN_TYPE = "6e4d7565-b277-4dfa-8b7d-fbf306d9d0cd";
@@ -118,10 +117,6 @@ public class TestConstants {
 
   public static String inventoryBoundWithTopic(String tenantId) {
     return getTopicName(tenantId, INVENTORY_BOUND_WITH_TOPIC);
-  }
-
-  public static String bibframeTopic(String tenantId) {
-    return getTopicName(tenantId, BIBFRAME_TOPIC);
   }
 
   public static String indexName(String tenantId) {
