+++ src/test/resources/application.yml
@@ -110,9 +110,6 @@ folio:
       - name: inventory.location
         numPartitions: 1
         replicationFactor: 1
-      - name: search.bibframe
-        numPartitions: 1
-        replicationFactor: 1
     listener:
       events:
         concurrency: 2
@@ -142,10 +139,6 @@ folio:
         concurrency: 1
         topic-pattern: (${folio.environment}\.)(.*\.)inventory\.location
         group-id: ${folio.environment}-mod-search-location-type-group
-      bibframe:
-        concurrency: ${KAFKA_BIBFRAME_CONCURRENCY:1}
-        topic-pattern: ${KAFKA_BIBFRAME_CONSUMER_PATTERN:(${folio.environment}\.)(.*\.)search\.bibframe}
-        group-id: ${folio.environment}-mod-search-bibframe-group
   okapiUrl: ${okapi.url}
   logging:
     request:
